Create Routine Form

+++ app/src/main/java/com/minizuure/todoplannereducationedition/services/database/routine/RoutineTableDao.kt
@@ -12,7 +12,7 @@ interface RoutineTableDao : BaseIODao<RoutineTable> {
     suspend fun getAll(): List<RoutineTable>
 
     @Query("SELECT * FROM routine_table WHERE id = :id")
-    suspend fun getById(id: Int): RoutineTable?
+    suspend fun getById(id: Long): RoutineTable?
 
     @Query("SELECT * FROM routine_table WHERE title LIKE :searchQuery LIMIT :limit OFFSET :offset")
     suspend fun getPaginated(limit: Int, offset: Int = 0, searchQuery: String = ""): List<RoutineTable>
